Move StateSpaceUtil into wpilibStateSpace
